build: enable -Wall -Wextra (icx) and zero all resulting warnings
Turn on -Wall -Wextra in the shared SYCL flag helper so it covers the infra lib
and every executable. Mark third-party include dirs (miniaudio, imgui, Level Zero)
SYSTEM and silence vendored imgui's own warnings so only our code is checked, then
fix every warning it surfaced:
- unused params (AsyncAudioPipeline::Start) and unused vars (benchmark_harness
  conv_channels, test_ui_manager 'state' under NDEBUG).
- missing-field-initializers on Level Zero / NOTIFYICONDATAW structs (value-init
  then set the type/size field).
- sign-compare in sycl_test_harness.
- getenv deprecation in test_kernel_correctness (localized clang pragma).
Build is warning-clean under -Wall -Wextra.

+++ tests/sycl_test_harness.h
@@ -44,7 +44,7 @@ public:
         }
         fprintf(stdout, "--- Result: %d/%zu tests passed ---\n", passed, m_tests.size());
         fflush(stdout);
-        return (passed == m_tests.size()) ? 0 : 1;
+        return (static_cast<size_t>(passed) == m_tests.size()) ? 0 : 1;
     }
 
 private:
